introduce DisplayModelConverter and use it in SVGConverter
